spark/requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.9
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
python-dotenv==1.0.0